+++ Echo/src/Core/Compute/Engine/Operation.cs
