+++ Echo/src/Core/Compute/Engine/OperationAbortedException.cs
